DashboardScene: Move DashboardGridItem
