+++ public/app/features/dashboard-scene/panel-edit/PanelEditor.test.ts
@@ -18,7 +18,7 @@ import { mockDataSource, MockDataSourceSrv } from 'app/features/alerting/unified
 import { DataSourceType } from 'app/features/alerting/unified/utils/datasource';
 import * as libAPI from 'app/features/library-panels/state/api';
 
-import { DashboardGridItem } from '../scene/DashboardGridItem';
+import { DashboardGridItem } from '../scene/layout-default/DashboardGridItem';
 import { DashboardScene } from '../scene/DashboardScene';
 import { LibraryPanelBehavior } from '../scene/LibraryPanelBehavior';
 import { DefaultGridLayoutManager } from '../scene/layout-default/DefaultGridLayoutManager';
