+++ public/app/features/dashboard-scene/scene/PanelSearchLayout.tsx
@@ -9,7 +9,7 @@ import { Trans } from 'app/core/internationalization';
 
 import { activateSceneObjectAndParentTree } from '../utils/utils';
 
-import { DashboardGridItem } from './DashboardGridItem';
+import { DashboardGridItem } from './layout-default/DashboardGridItem';
 import { DashboardScene } from './DashboardScene';
 import { DefaultGridLayoutManager } from './layout-default/DefaultGridLayoutManager';
