+++ public/app/features/dashboard-scene/scene/ViewPanelScene.test.tsx
@@ -1,6 +1,6 @@
 import { LocalValueVariable, SceneGridLayout, SceneGridRow, SceneVariableSet, VizPanel } from '@grafana/scenes';
 
-import { DashboardGridItem } from './DashboardGridItem';
+import { DashboardGridItem } from './layout-default/DashboardGridItem';
 import { DashboardScene } from './DashboardScene';
 import { ViewPanelScene } from './ViewPanelScene';
 import { DefaultGridLayoutManager } from './layout-default/DefaultGridLayoutManager';
