+++ public/app/features/dashboard-scene/scene/layout-default/DefaultGridLayoutManager.test.tsx
@@ -1,7 +1,7 @@
 import { SceneGridItemLike, SceneGridLayout, SceneGridRow, SceneQueryRunner, VizPanel } from '@grafana/scenes';
 
 import { findVizPanelByKey } from '../../utils/utils';
-import { DashboardGridItem } from '../DashboardGridItem';
+import { DashboardGridItem } from './DashboardGridItem';
 
 import { DefaultGridLayoutManager } from './DefaultGridLayoutManager';
