+++ public/app/features/dashboard-scene/scene/row-actions/RowActions.tsx
@@ -15,7 +15,7 @@ import { SHARED_DASHBOARD_QUERY } from 'app/plugins/datasource/dashboard';
 import { ShowConfirmModalEvent } from 'app/types/events';
 
 import { getDashboardSceneFor, getQueryRunnerFor } from '../../utils/utils';
-import { DashboardGridItem } from '../DashboardGridItem';
+import { DashboardGridItem } from '../layout-default/DashboardGridItem';
 import { DashboardScene } from '../DashboardScene';
 import { RowRepeaterBehavior } from '../RowRepeaterBehavior';
 import { DefaultGridLayoutManager } from '../layout-default/DefaultGridLayoutManager';
