+++ public/app/features/dashboard-scene/sharing/ShareLibraryPanelTab.tsx
@@ -6,7 +6,7 @@ import { ShareLibraryPanel } from 'app/features/dashboard/components/ShareModal/
 import { shareDashboardType } from 'app/features/dashboard/components/ShareModal/utils';
 import { DashboardModel, PanelModel } from 'app/features/dashboard/state';
 
-import { DashboardGridItem } from '../scene/DashboardGridItem';
+import { DashboardGridItem } from '../scene/layout-default/DashboardGridItem';
 import { gridItemToPanel, transformSceneToSaveModel } from '../serialization/transformSceneToSaveModel';
 import { getDashboardSceneFor } from '../utils/utils';
