+++ public/app/features/dashboard-scene/utils/dashboardSceneGraph.test.ts
@@ -4,7 +4,7 @@ import { DashboardCursorSync } from '@grafana/schema';
 import { DashboardAnnotationsDataLayer } from '../scene/DashboardAnnotationsDataLayer';
 import { DashboardControls } from '../scene/DashboardControls';
 import { DashboardDataLayerSet } from '../scene/DashboardDataLayerSet';
-import { DashboardGridItem } from '../scene/DashboardGridItem';
+import { DashboardGridItem } from '../scene/layout-default/DashboardGridItem';
 import { DashboardScene, DashboardSceneState } from '../scene/DashboardScene';
 import { VizPanelLinks, VizPanelLinksMenu } from '../scene/PanelLinks';
 import { DefaultGridLayoutManager } from '../scene/layout-default/DefaultGridLayoutManager';
