+++ public/app/features/library-panels/state/api.ts
@@ -3,7 +3,7 @@ import { lastValueFrom } from 'rxjs';
 import { VizPanel } from '@grafana/scenes';
 import { LibraryPanel, defaultDashboard } from '@grafana/schema';
 import { DashboardModel } from 'app/features/dashboard/state';
-import { DashboardGridItem } from 'app/features/dashboard-scene/scene/DashboardGridItem';
+import { DashboardGridItem } from 'app/features/dashboard-scene/scene/layout-default/DashboardGridItem';
 import { vizPanelToPanel } from 'app/features/dashboard-scene/serialization/transformSceneToSaveModel';
 import { getLibraryPanelBehavior } from 'app/features/dashboard-scene/utils/utils';
